Renamed Limit::__invoke() to Limit::hasBeenReached(), step 1
The use of objects which are callables lead to non-intuitive behavior of Dependency Injection Containers like Pimple:
https://github.com/nlisgo/annotations-silex/commit/8e8e584422c7fb7e94f26e4b2e707420a11204bd
To avoid the corner cases, we can transform them dropping the callable interface. We were already going into this direction when introducing the `Limit` interface and enforcing a type hint stricter than callable.